admin_service/requirements.txt
Django>=5.0
djangorestframework
gunicorn
mysqlclient>=2.2
django-cors-headers==4.4.0